% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decide_happy_hour.R
\name{decide_happy_hour}
\alias{decide_happy_hour}
\title{Decide the best day for an happy hour}
\usage{
decide_happy_hour(x)
}
\arguments{
\item{x}{A list of Dates vectors}
}
\value{
the set of common date(s)
}
\description{
Decide the best day for an happy hour
}
\examples{
1+1
}
